# Calculate luminances up front for efficiency.