This changes the luminance calculation to be done once up front rather than repeatedly during the energy calculation pass. This improves running time at the cost of more memory usage.

Before:

```
> go test -run=XXX -bench=EnergySummation -count=8
BenchmarkEnergySummation-56    	      60	  19912731 ns/op
BenchmarkEnergySummation-56    	      58	  19881756 ns/op
BenchmarkEnergySummation-56    	      60	  20051701 ns/op
BenchmarkEnergySummation-56    	      57	  19925334 ns/op
BenchmarkEnergySummation-56    	      60	  19934364 ns/op
BenchmarkEnergySummation-56    	      60	  19991013 ns/op
BenchmarkEnergySummation-56    	      58	  20027357 ns/op
BenchmarkEnergySummation-56    	      58	  19939651 ns/op
```

After:

```
> go test -run=XXX -bench=EnergySummation -count=8
BenchmarkEnergySummation-56    	     214	   5544896 ns/op
BenchmarkEnergySummation-56    	     217	   5530105 ns/op
BenchmarkEnergySummation-56    	     217	   5513593 ns/op
BenchmarkEnergySummation-56    	     217	   5521769 ns/op
BenchmarkEnergySummation-56    	     216	   5518241 ns/op
BenchmarkEnergySummation-56    	     218	   5512376 ns/op
BenchmarkEnergySummation-56    	     218	   5503996 ns/op
BenchmarkEnergySummation-56    	     217	   5493456 ns/op
```